+++ shaderlib/project/index.js
@@ -1,8 +1,0 @@
-/* eslint-disable */
-var fs = require('fs');
-module.exports = {
-  project: {
-  	interface: 'project',
-  	source: fs.readFileSync(__dirname + '/project.glsl', 'utf8')
-  }
-};