+++ src/lib/layer-manager.js
@@ -28,10 +28,10 @@
 //
 /* eslint-disable no-try-catch */
 import Layer from './layer';
-import {Viewport} from '../viewport';
 import {log} from './utils';
 import assert from 'assert';
-import {pickLayers} from './pick-layers';
+import {drawLayers, pickLayers} from './draw-and-pick';
+import {Viewport} from 'viewport-mercator-project';
 import {FramebufferObject} from 'luma.gl';
 
 export default class LayerManager {
@@ -53,34 +53,17 @@ export default class LayerManager {
     Object.seal(this.context);
   }
 
-  setContext({
-    width, height, latitude, longitude, zoom, pitch, bearing, altitude
-  }) {
-    /* eslint-disable */
+  setViewport(viewport) {
+    assert(viewport instanceof Viewport, 'Invalid viewport');
     const oldViewport = this.context.viewport;
-    const viewportChanged = !oldViewport ||
-      width !== oldViewport.width ||
-      height !== oldViewport.height ||
-      latitude !== oldViewport.latitude ||
-      longitude !== oldViewport.longitude ||
-      zoom !== oldViewport.zoom ||
-      bearing !== oldViewport.bearing ||
-      pitch !== oldViewport.pitch ||
-      altitude !== oldViewport.altitude;
+    const viewportChanged = !oldViewport || !viewport.equals(oldViewport);
 
-    if (viewportChanged || !this.context.viewport) {
+    if (viewportChanged) {
       Object.assign(this.oldContext, this.context);
-
-      const viewport = new Viewport({
-        width, height, latitude, longitude, zoom, pitch, bearing, altitude,
-        tileSize: 512
-      });
-
       this.context.viewport = viewport;
-      this.context.viewportChanged = viewportChanged;
+      this.context.viewportChanged = true;
       this.context.uniforms = {};
-
-      log(4, viewport, latitude, longitude, zoom);
+      log(4, viewport);
     }
 
     return this;
@@ -115,14 +98,7 @@ export default class LayerManager {
   drawLayers() {
     assert(this.context.viewport, 'LayerManager.drawLayers: viewport not set');
 
-    const {uniforms} = this.context;
-    let layerIndex = 0;
-    for (const layer of this.layers) {
-      if (layer.props.visible) {
-        layer.drawLayer({uniforms, layerIndex});
-        layerIndex++;
-      }
-    }
+    drawLayers({layers: this.layers});
 
     return this;
   }
@@ -132,8 +108,8 @@ export default class LayerManager {
 
     // Set up a frame buffer if needed
     if (this.context.pickingFBO === null ||
-    gl.canvas.width !== this.context.pickingFBO.width ||
-    gl.canvas.height !== this.context.pickingFBO.height) {
+      gl.canvas.width !== this.context.pickingFBO.width ||
+      gl.canvas.height !== this.context.pickingFBO.height) {
       this.context.pickingFBO = new FramebufferObject(gl, {
         width: gl.canvas.width,
         height: gl.canvas.height
@@ -142,7 +118,10 @@ export default class LayerManager {
     return pickLayers(gl, {
       x,
       y,
-      uniforms,
+      uniforms: {
+        renderPickingBuffer: true,
+        picking_uEnable: true
+      },
       layers: this.layers,
       mode,
       pickingFBO: this.context.pickingFBO
@@ -258,10 +237,8 @@ export default class LayerManager {
     const {state, props} = oldLayer;
 
     // sanity check
-    assert(state,
-      'deck.gl sanity check - Matching layer has no state');
-    assert(oldLayer !== newLayer,
-      'deck.gl sanity check - Matching layer is same');
+    assert(state, 'deck.gl sanity check - Matching layer has no state');
+    assert(oldLayer !== newLayer, 'deck.gl sanity check - Matching layer is same');
 
     // Move state
     newLayer.state = state;
@@ -303,9 +280,7 @@ export default class LayerManager {
           changeFlags: layer.diffProps({}, layer.props, this.context)
         });
       } catch (err) {
-        log.once(0,
-          `deck.gl error during initialization of ${layerName(layer)} ${err}`,
-          err);
+        log.once(0, `deck.gl error during initialization of ${layerName(layer)} ${err}`, err);
         // Save first error
         error = error || err;
       }
@@ -336,8 +311,7 @@ export default class LayerManager {
           changeFlags: layer.diffProps(oldProps, layer.props, this.context)
         });
       } catch (err) {
-        log.once(0,
-          `deck.gl error during update of ${layerName(layer)}`, err);
+        log.once(0, `deck.gl error during update of ${layerName(layer)}`, err);
         // Save first error
         error = err;
       }