+++ src/react/webgl-renderer.js
@@ -20,7 +20,7 @@
 
 /* global window */
 import React, {PropTypes} from 'react';
-import autobind from 'autobind-decorator';
+import autobind from 'react-autobind';
 import {createGLContext} from 'luma.gl';
 /* global requestAnimationFrame, cancelAnimationFrame */
 
@@ -38,7 +38,6 @@ export default class WebGLRenderer extends React.Component {
 
     pixelRatio: PropTypes.number,
     viewport: PropTypes.object.isRequired,
-    blending: PropTypes.object,
     events: PropTypes.object,
     gl: PropTypes.object,
     glOptions: PropTypes.object,
@@ -85,6 +84,7 @@ export default class WebGLRenderer extends React.Component {
     this.state = {};
     this._animationFrame = null;
     this.gl = null;
+    autobind(this);
   }
 
   componentDidMount() {
@@ -124,7 +124,6 @@ export default class WebGLRenderer extends React.Component {
   /**
    * Main WebGL animation loop
    */
-  @autobind
   _animationLoop() {
     this._renderFrame();
     // Keep registering ourselves for the next animation frame