+++ src/test/electron.js
@@ -1,5 +1,4 @@
 require('babel-polyfill');
 
-require('./');
-require('../experimental/test');
+require('./index');
 require('./react');