+++ src/test/headless.js
@@ -1,2 +1,3 @@
-require('./');
-require('../experimental/test');
+import 'babel-polyfill';
+import 'luma.gl/headless';
+import './index';