+++ src/test/lib/index.js
@@ -1,3 +1,4 @@
-import './layer.spec';
-import './attribute-manager.spec';
 import './utils/compare-objects.spec';
+import './attribute-manager.spec';
+import './viewport-uniforms.spec';
+import './layer.spec';